Extracted the codeclimate client
